examples/twilio-gemini-live/requirements.txt
pipecat-ai[google,websocket,silero]
aiofiles
python-dotenv
fastapi
uvicorn
loguru